refactor(webapp): unify checkbox/accordion/tabs (preserve-both) + relocate progress (A1)
Continue A1 primitive unification into the canonical L0 layer.

- checkbox, accordion, tabs: preserve-both — app look kept verbatim as the bare
  exports (app call-sites untouched), widget look relocated under `*Widget` names
  and aliased back by the widget shim. Shared identical parts (Accordion/Tabs =
  the Radix Root) are shared, not duplicated. All on leaf @/lib/cn.
- progress: widget-only clean primitive relocated to components/ui (shim left at
  the widget path).
- input: intentionally NOT relocated — it imports Tree-3 composites (HStack,
  Text, Warning, error tooltip), so it is a widget composite, not an L0
  primitive; out of scope for primitive unification (note for a re-home pass).

Verified: every app/widget section's class strings diff byte-identical vs HEAD;
tsc --noEmit green; 10-file widget/module batch green (79 tests); PSM isolation
canary still green (no app-graph coupling reintroduced).
