+++ apps/webapp/src/components/ui/checkbox.tsx
@@ -2,8 +2,10 @@ import * as React from 'react';
 import * as CheckboxPrimitive from '@radix-ui/react-checkbox';
 import { Check, Minus } from 'lucide-react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib/cn';
 
+/* App checkbox (canonical `Checkbox`) — original components/ui look. UNCHANGED.
+ * Superset of the widget look (supports the `indeterminate` state). */
 const Checkbox = React.forwardRef<
   React.ElementRef<typeof CheckboxPrimitive.Root>,
   React.ComponentPropsWithoutRef<typeof CheckboxPrimitive.Root>
@@ -31,4 +33,26 @@ const Checkbox = React.forwardRef<
 ));
 Checkbox.displayName = CheckboxPrimitive.Root.displayName;
 
-export { Checkbox };
+/* Widget checkbox (`CheckboxWidget`) — original widgets look, preserved verbatim
+ * per ticket A1's preserve-both decision (different border token, no
+ * indeterminate state). Exposed via the widget shim as `Checkbox`. */
+const CheckboxWidget = React.forwardRef<
+  React.ElementRef<typeof CheckboxPrimitive.Root>,
+  React.ComponentPropsWithoutRef<typeof CheckboxPrimitive.Root>
+>(({ className, ...props }, ref) => (
+  <CheckboxPrimitive.Root
+    ref={ref}
+    className={cn(
+      'border-textDesaturated ring-offset-background focus-visible:ring-ring data-[state=checked]:bg-radial-(--gradient-position) data-[state=checked]:from-primary-start/100 data-[state=checked]:to-primary-end/100 data-[state=checked]:text-primary-foreground focus-visible:outline-hidden peer h-4 w-4 shrink-0 rounded-sm border focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50',
+      className
+    )}
+    {...props}
+  >
+    <CheckboxPrimitive.Indicator className={cn('flex items-center justify-center text-current')}>
+      <Check className="h-4 w-4" />
+    </CheckboxPrimitive.Indicator>
+  </CheckboxPrimitive.Root>
+));
+CheckboxWidget.displayName = CheckboxPrimitive.Root.displayName;
+
+export { Checkbox, CheckboxWidget };
