+++ apps/webapp/src/widgets/components/ui/progress.tsx
@@ -1,23 +1,4 @@
-import * as React from 'react';
-import * as ProgressPrimitive from '@radix-ui/react-progress';
-
-import { cn } from '@/widgets/lib/utils';
-
-const Progress = React.forwardRef<
-  React.ElementRef<typeof ProgressPrimitive.Root>,
-  React.ComponentPropsWithoutRef<typeof ProgressPrimitive.Root>
->(({ className, value, ...props }, ref) => (
-  <ProgressPrimitive.Root
-    ref={ref}
-    className={cn('relative h-4 w-full overflow-hidden rounded-full bg-white/40', className)}
-    {...props}
-  >
-    <ProgressPrimitive.Indicator
-      className="h-full w-full flex-1 bg-white/60 transition-all"
-      style={{ transform: `translateX(-${100 - (value || 0)}%)` }}
-    />
-  </ProgressPrimitive.Root>
-));
-Progress.displayName = ProgressPrimitive.Root.displayName;
-
-export { Progress };
+// Re-export shim — unified into the canonical L0 primitive (ticket A1).
+// Implementation lives in @/components/ui/progress; this path is kept so existing
+// widget call-sites keep compiling. Slated for removal once call-sites migrate.
+export * from '@/components/ui/progress';
